For coverage action only fetch origin/master
This change is to explicitly fetch origin/master for the coverage comparison. The original git fetch here was based on a misunderstanding (or possibly a change) to how the GitHub checkout action works, and was attempting to ensure that master and the PR branches were all available. That turns out to not be necessary.